requirements.txt
# Core ML/AI
torch>=2.3.0
transformers>=4.36.0
sentence-transformers>=2.2.2
einops>=0.6.0
timm>=0.9.0

# Vector Database
chromadb>=0.4.0

# LLM Serving
vllm>=0.6.0
flash-attn>=2.6.0
bitsandbytes>=0.44.0

# Web Search
ddgs>=7.0.0
beautifulsoup4>=4.12.0

# Web Interface
gradio>=4.0.0
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Optional for better performance
accelerate>=0.25.0

# Advanced OCR (optional - install separately if needed)
# paddlepaddle==2.6.0
# paddleocr>=2.7.0
# easyocr>=1.7.0
# pix2tex>=0.1.2

# Document Understanding (optional - install separately if needed)
# layoutparser[layoutmodels,tesseract]>=0.3.4
# pdf2image>=1.16.0

# Math Processing
sympy>=1.12
matplotlib>=3.7.0

# Enhanced NLP
openai>=1.0.0
anthropic>=0.8.0
tiktoken>=0.5.0